Upload of final version of project
Contains all packages necessary for replication of project, excluding one data file which was too large to upload. This file was not utilized in the making of the paper, but is part of the original replication package from Nuzzo, Jennifer B., and Jorge R. Ledesma (2023) and was included to maintain consistency and accuracy to everything used in the creation of this project. File in question is "UN_pops.csv" and can be retrieved at https://www.openicpsr.org/openicpsr/project/193684/version/V1/view?path=/openicpsr/193684/fcr:versions/V1&type=project